// src/webparts/travelSupportBrasil/Interfaces/IRequests.ts
export interface IRequests {
  Id?: number;
  MACROPROCESSO: string;
  PROCESSO: string;
  WF_APROVACAO: boolean;
  ALCADA_APROVACAO?: string;
  SLA: number;
  AREA_RESOLVEDORA: string;
  APROVADOR_ID?: string;
  APROVADOR_NOME?: string;
  APROVADOR_EMAIL?: string;
  GESTOR_DA_BASE_ID?: string;
  GESTOR_DA_BASE_NOME?: string;
  GESTOR_DA_BASE_EMAIL?: string;
  BENEFICIARIO_ID?: string;
  BENEFICIARIO_NOME?: string;
  BENEFICIARIO_EMAIL?: string;
  CENTRO_DE_CUSTOS?: string;
  CNPJ_DE_FATURAMENTO?: string;
  NOVO_LIMITE?: string;
  QUANTIDADE_DE_CARTOES?: string;
  CODIGO_DA_BASE?: string;
  MOTIVO?: string;
  ULTIMO_PERIODO_DA_MEDICAO?: string;
  PERIODO?: string;
  ESTABELECIMENTO?: string;
  DATA_DE_UTILIZACAO?: Date;
  CODIGO_DO_RAMO_DE_ATIVIDADE?: string;
  CPF?: string;
  TELEFONE?: string;
  TIPO_DE_SOLICITACAO?: string;
  DONO_DA_DESPESA_ID?: string;
  DONO_DA_DESPESA_NOME?: string;
  DONO_DA_DESPESA_EMAIL?: string;
  TIPO_DE_DELEGACAO?: string;
  RELATORIO_CONCUR?: string;
  EMPREGADO_ATIVO?: boolean;
  VALOR?: number;
  DATA_DO_DESCONTO?: Date;
  COD_DO_DESCONTO_NO_CONTRACHEQUE?: string;
  NACIONAL_INTERNACIONAL?: string;
  ACOMPANHANTES?: string;
  BENEFICIARIO_EMPRESA_COD?: string;
  GESTOR_DA_BASE_EMPRESA_COD?: string;
  APROVADOR_EMPRESA_COD?: string;
  DONO_DA_DESPESA_EMPRESA_COD?: string;
  ULTIMOS_DIGITOS_DO_CARTAO?: string;
  DATA_DE_APROVACAO?: Date;
  STATUS_APROVACAO?: string;
  STATUS_ATENDIMENTO?: string;
  DATA_DE_ATENDIMENTO?: Date;
  APROVACAO_COMENTARIOS?: string;
  ATENDIMENTO_COMENTARIOS?: string;
  APROVADOR_LEVEL?: string;
  BENEFICIARIO_EMPRESA_NOME?: string;
  APROVADOR_EMPRESA_NOME?: string;
  DONO_DA_DESPESA_EMPRESA_NOME?: string;
  GESTOR_DA_BASE_EMPRESA_NOME?: string;
  RESP_MEDICAO_EMPRESA_NOME?: string;
  END_CEP?: string;
  END_LOGRADOURO?: string;
  END_NUMERO?: number;
  END_COMPLEMENTO?: string;
  RESP_MEDICAO_EMPRESA_COD?: string;
  RESP_MEDICAO_ID?: string;
  RESP_MEDICAO_NOME?: string;
  RESP_MEDICAO_EMAIL?: string;
}